requirements.txt
langgraph
langgraph-cli
langchain
langchain-community
faiss-cpu
pydantic
python-dotenv
pytest
openai>=1.6.0
langchain-openai>=0.2.0
streamlit
IPython
graphviz
pygraphviz